update: remove unused libs, code unnecessary
