--- src/main/java/com/github/kaivu/common/utils/ImageUtil.java
@@ -1,125 +1,0 @@
-package com.github.kaivu.common.utils;
-
-import lombok.extern.slf4j.Slf4j;
-
-import java.awt.*;
-import java.awt.geom.Area;
-import java.awt.geom.RoundRectangle2D;
-import java.awt.image.BufferedImage;
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.net.URL;
-
-import javax.imageio.ImageIO;
-
-/**
- * Created by Khoa Vu.
- * Mail: khoavd12@fpt.com
- * Date: 3/15/25
- * Time: 1:13 AM
- */
-@Slf4j
-public class ImageUtil {
-
-    private ImageUtil() {
-        throw new IllegalStateException("Utility class");
-    }
-
-    private static Image getScaledImage(BufferedImage img, int maxSize, int type) {
-        int w0 = img.getWidth();
-        int h0 = img.getHeight();
-        int w = w0;
-        int h = h0;
-        if (type == 1) {
-            w = w0 > h0 ? maxSize : (maxSize * w0 / h0);
-            h = w0 > h0 ? (maxSize * h0 / w0) : maxSize;
-        } else if (type == 2) {
-            w = w0 > h0 ? (maxSize * w0 / h0) : maxSize;
-            h = w0 > h0 ? maxSize : (maxSize * h0 / w0);
-        }
-        Image image = img.getScaledInstance(w, h, Image.SCALE_SMOOTH);
-        BufferedImage result = new BufferedImage(w, h, BufferedImage.TYPE_INT_ARGB);
-        Graphics2D g = result.createGraphics();
-        g.drawImage(image, 0, 0, null);
-        // Draw at the appropriate position
-        return result;
-    }
-
-    private static BufferedImage getRoundedImage(BufferedImage img, int size) {
-        return getRoundedImage(img, size, size / 2, 2);
-    }
-
-    private static BufferedImage getRoundedImage(BufferedImage img, int size, int radius, int type) {
-        BufferedImage result = new BufferedImage(size, size, BufferedImage.TYPE_INT_ARGB);
-        Graphics2D g = result.createGraphics();
-        // First scale proportionally with minimum width/height as size, then center the image and cut out a circular
-        // image with diameter of size
-        Image fixedImg = getScaledImage(img, size, type);
-        g.drawImage(fixedImg, (size - fixedImg.getWidth(null)) / 2, (size - fixedImg.getHeight(null)) / 2, null);
-        // Draw at the appropriate position
-        // Rounded corners
-        if (radius > 0) {
-            RoundRectangle2D round = new RoundRectangle2D.Double(0, 0, size, size, radius * 2, radius * 2);
-            Area clear = new Area(new Rectangle(0, 0, size, size));
-            clear.subtract(new Area(round));
-            g.setComposite(AlphaComposite.Clear);
-            // Anti-aliasing
-            g.setRenderingHint(RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON);
-            g.fill(clear);
-            g.dispose();
-        }
-        return result;
-    }
-
-    protected static BufferedImage get24BitImage(BufferedImage $image) {
-        int __w = $image.getWidth();
-        int __h = $image.getHeight();
-        int[] __imgARGB = getRGBs($image.getRGB(0, 0, __w, __h, null, 0, __w));
-        BufferedImage __newImg = new BufferedImage(__w, __h, BufferedImage.TYPE_INT_RGB);
-        __newImg.setRGB(0, 0, __w, __h, __imgARGB, 0, __w);
-        return __newImg;
-    }
-
-    protected static BufferedImage get24BitImage(BufferedImage $image, Color $bgColor) {
-        int $w = $image.getWidth();
-        int $h = $image.getHeight();
-        BufferedImage img = new BufferedImage($w, $h, BufferedImage.TYPE_INT_RGB);
-        Graphics2D g = img.createGraphics();
-        g.setColor($bgColor);
-        g.fillRect(0, 0, $w, $h);
-        g.drawRenderedImage($image, null);
-        g.dispose();
-        return img;
-    }
-
-    public static int[] getRGBs(int[] $argb) {
-        int[] __rgbs = new int[$argb.length];
-        for (int i = 0; i < $argb.length; i++) {
-            __rgbs[i] = $argb[i] & 0xFFFFFF;
-        }
-        return __rgbs;
-    }
-
-    public static void toPNG(File img, File save, int size, int radius) throws IOException {
-        ImageIO.write(getRoundedImage(ImageIO.read(img), size, radius, 2), "PNG", save);
-    }
-
-    public static ByteArrayOutputStream toPNG(URL url, int size, int radius) throws IOException {
-        ByteArrayOutputStream os = new ByteArrayOutputStream();
-        ImageIO.write(getRoundedImage(ImageIO.read(url), size, radius, 2), "PNG", os);
-        return os;
-    }
-
-    public static void toPNG(URL url, OutputStream os, int size, int radius) throws IOException {
-        BufferedImage roundedImage = getRoundedImage(ImageIO.read(url), size, radius, 2);
-        ImageIO.write(roundedImage, "PNG", os);
-        os.close();
-    }
-
-    private static String getTempPrefix(String key) {
-        String tempPrefix = "headImage";
-        return tempPrefix + "_" + key;
-    }
-}
